QtQml: Allow more fine grained control of the disk cache
You can now enable and disable the AOT-compiled code and the loading and
saving of .qmlc files separately.

Fixes: QTBUG-101358

+++ tests/auto/qml/qmlcachegen/tst_qmlcachegen.cpp
@@ -120,7 +120,9 @@ tst_qmlcachegen::tst_qmlcachegen()
 
 void tst_qmlcachegen::initTestCase()
 {
-    qputenv("QML_FORCE_DISK_CACHE", "1");
+    if (qEnvironmentVariableIsEmpty("QML_DISK_CACHE"))
+        qputenv("QML_FORCE_DISK_CACHE", "1");
+
     QStandardPaths::setTestModeEnabled(true);
 
     // make sure there's no pre-existing cache dir
@@ -131,29 +133,54 @@ void tst_qmlcachegen::initTestCase()
     QQmlDataTest::initTestCase();
 }
 
+#if QT_CONFIG(process)
+static void testWithEnvironment(
+        const QString &function, const QString &testFilePath, const QByteArray &value, bool success)
+{
+    QProcess child;
+    child.setProgram(QCoreApplication::applicationFilePath());
+    child.setArguments(QStringList(function));
+    QProcessEnvironment env = QProcessEnvironment::systemEnvironment();
+    env.remove(QLatin1String("QML_FORCE_DISK_CACHE"));
+    env.insert(QLatin1String("QMLCACHEGEN_TEST_FILE_PATH"), testFilePath);
+    env.insert(QLatin1String("QML_DISK_CACHE"), QLatin1String(value));
+    child.setProcessEnvironment(env);
+    child.start();
+    QVERIFY(child.waitForFinished());
+    if (success)
+        QCOMPARE(child.exitCode(), 0);
+    else
+        QVERIFY(child.exitCode() != 0);
+}
+#endif
+
 void tst_qmlcachegen::loadGeneratedFile()
 {
 #if defined(QTEST_CROSS_COMPILED)
     QSKIP("Cannot call qmlcachegen on cross-compiled target.");
 #endif
-
     QTemporaryDir tempDir;
     QVERIFY(tempDir.isValid());
 
-    const auto writeTempFile = [&tempDir](const QString &fileName, const char *contents) {
-        QFile f(tempDir.path() + '/' + fileName);
-        const bool ok = f.open(QIODevice::WriteOnly | QIODevice::Truncate);
-        Q_ASSERT(ok);
-        f.write(contents);
-        return f.fileName();
-    };
+    const QByteArray path = qgetenv("QMLCACHEGEN_TEST_FILE_PATH");
+    QString testFilePath;
+    if (path.isEmpty()) {
+        const auto writeTempFile = [&tempDir](const QString &fileName, const char *contents) {
+            QFile f(tempDir.path() + '/' + fileName);
+            const bool ok = f.open(QIODevice::WriteOnly | QIODevice::Truncate);
+            Q_ASSERT(ok);
+            f.write(contents);
+            return f.fileName();
+        };
 
-    const QString testFilePath = writeTempFile("test.qml", "import QtQml 2.0\n"
-                                                           "QtObject {\n"
-                                                           "    property int value: Math.min(100, 42);\n"
-                                                           "}");
-
-    QVERIFY(generateCache(testFilePath));
+        testFilePath = writeTempFile("test.qml", "import QtQml 2.0\n"
+                                                 "QtObject {\n"
+                                                 "    property int value: Math.min(100, 42);\n"
+                                                 "}");
+        QVERIFY(generateCache(testFilePath));
+    } else {
+        testFilePath = QString::fromUtf8(path);
+    }
 
     const QString cacheFilePath = testFilePath + QLatin1Char('c');
     QVERIFY(QFile::exists(cacheFilePath));
@@ -168,7 +195,17 @@ void tst_qmlcachegen::loadGeneratedFile()
         QCOMPARE(uint(cacheUnit->sourceFileIndex), uint(0));
     }
 
-    QVERIFY(QFile::remove(testFilePath));
+    if (path.isEmpty()) {
+        QVERIFY(QFile::remove(testFilePath));
+    } else {
+#if QT_CONFIG(process)
+        testWithEnvironment("loadGeneratedFile", testFilePath, "qmlc-read", true);
+        testWithEnvironment("loadGeneratedFile", testFilePath, "qmlc-write", false);
+        testWithEnvironment("loadGeneratedFile", testFilePath, "qmlc-read,qmlc-read", true);
+        testWithEnvironment("loadGeneratedFile", testFilePath, "qmlc", true);
+        testWithEnvironment("loadGeneratedFile", testFilePath, "wrong", false);
+#endif
+    }
 
     QQmlEngine engine;
     CleanlyLoadingComponent component(&engine, QUrl::fromLocalFile(testFilePath));
@@ -394,6 +431,8 @@ static const char *versionCheckErrorString(QQmlMetaType::CachedUnitLookupError e
         return "no unit found";
     case QQmlMetaType::CachedUnitLookupError::VersionMismatch:
         return "version mismatch";
+    case QQmlMetaType::CachedUnitLookupError::NotFullyTyped:
+        return "unit not fully typed";
     }
 
     return "wat?";
@@ -408,7 +447,7 @@ void tst_qmlcachegen::versionChecksForAheadOfTimeUnits()
     QQmlMetaType::CachedUnitLookupError error = QQmlMetaType::CachedUnitLookupError::NoError;
     QLoggingCategory::setFilterRules("qt.qml.diskcache.debug=true");
     const QQmlPrivate::CachedQmlUnit *originalUnit = QQmlMetaType::findCachedCompilationUnit(
-            QUrl("qrc:/data/versionchecks.qml"), &error);
+            QUrl("qrc:/data/versionchecks.qml"), QQmlMetaType::AcceptUntyped, &error);
     QLoggingCategory::setFilterRules(QString());
     QVERIFY2(originalUnit, versionCheckErrorString(error));
     QV4::CompiledData::Unit *tweakedUnit = (QV4::CompiledData::Unit *)malloc(originalUnit->qmlData->unitSize);
@@ -437,7 +476,8 @@ void tst_qmlcachegen::versionChecksForAheadOfTimeUnits()
 
     {
         QQmlMetaType::CachedUnitLookupError error = QQmlMetaType::CachedUnitLookupError::NoError;
-        QVERIFY(!QQmlMetaType::findCachedCompilationUnit(QUrl("qrc:/data/versionchecks.qml"), &error));
+        QVERIFY(!QQmlMetaType::findCachedCompilationUnit(
+                    QUrl("qrc:/data/versionchecks.qml"), QQmlMetaType::AcceptUntyped, &error));
         QCOMPARE(error, QQmlMetaType::CachedUnitLookupError::VersionMismatch);
     }
 
@@ -463,7 +503,7 @@ void tst_qmlcachegen::skippedResources()
 
     QQmlMetaType::CachedUnitLookupError error = QQmlMetaType::CachedUnitLookupError::NoError;
     const QQmlPrivate::CachedQmlUnit *unit = QQmlMetaType::findCachedCompilationUnit(
-            QUrl("qrc:/not/Skip.qml"), &error);
+            QUrl("qrc:/not/Skip.qml"), QQmlMetaType::AcceptUntyped, &error);
     QCOMPARE(unit, nullptr);
     QCOMPARE(error, QQmlMetaType::CachedUnitLookupError::NoUnitFound);
 }
@@ -581,6 +621,23 @@ void tst_qmlcachegen::qrcScriptImport()
     QScopedPointer<QObject> obj(component.create());
     QVERIFY(!obj.isNull());
     QTRY_COMPARE(obj->property("value").toInt(), 42);
+#if QT_CONFIG(process)
+    if (qEnvironmentVariableIsEmpty("QMLCACHEGEN_TEST_FILE_PATH")) {
+        testWithEnvironment("qrcScriptImport", ":/data/jsimport.qml", "aot-native", false);
+        testWithEnvironment("qrcScriptImport", ":/data/jsimport.qml", "aot-bytecode", true);
+        testWithEnvironment("qrcScriptImport", ":/data/jsimport.qml", "aot-bytecode,aot-native", true);
+        testWithEnvironment("qrcScriptImport", ":/data/jsimport.qml", "aot", true);
+        testWithEnvironment("qrcScriptImport", ":/data/jsimport.qml", "wrong", false);
+    }
+#endif
+
+    auto componentPrivate = QQmlComponentPrivate::get(&component);
+    QVERIFY(componentPrivate);
+    auto compilationUnit = componentPrivate->compilationUnit;
+    QVERIFY(compilationUnit);
+    auto unitData = compilationUnit->unitData();
+    QVERIFY(unitData);
+    QVERIFY(unitData->flags & QV4::CompiledData::Unit::StaticData);
 }
 
 void tst_qmlcachegen::fsScriptImport()
@@ -666,7 +723,7 @@ void tst_qmlcachegen::moduleScriptImport()
 
         QQmlMetaType::CachedUnitLookupError error = QQmlMetaType::CachedUnitLookupError::NoError;
         const QQmlPrivate::CachedQmlUnit *unitFromResources = QQmlMetaType::findCachedCompilationUnit(
-                QUrl("qrc:/data/script.mjs"), &error);
+                QUrl("qrc:/data/script.mjs"), QQmlMetaType::AcceptUntyped, &error);
         QVERIFY(unitFromResources);
 
         QCOMPARE(unitFromResources->qmlData, compilationUnit->unitData());
@@ -697,7 +754,7 @@ void tst_qmlcachegen::sourceFileIndices()
 
     QQmlMetaType::CachedUnitLookupError error = QQmlMetaType::CachedUnitLookupError::NoError;
     const QQmlPrivate::CachedQmlUnit *unitFromResources = QQmlMetaType::findCachedCompilationUnit(
-            QUrl("qrc:/data/versionchecks.qml"), &error);
+            QUrl("qrc:/data/versionchecks.qml"), QQmlMetaType::AcceptUntyped, &error);
     QVERIFY(unitFromResources);
     QVERIFY(unitFromResources->qmlData->flags & QV4::CompiledData::Unit::PendingTypeCompilation);
     QCOMPARE(uint(unitFromResources->qmlData->sourceFileIndex), uint(0));
